feat(Combinatorics/Additive): randomising by a function of dissociated support (#18916)
Prove that a function from a finite abelian group can be randomised by a function of
dissociated support.

Precisely, for `G` a finite abelian group and two functions `c : Dual G → ℝ` and `d : Dual G → ℝ`
such that `{ψ | d ψ ≠ 0} is dissociated`, the product of the `c ψ` over `ψ` is the same as the
average over `a` of the product of the `c ψ + Re (d ψ * ψ a)`.

From LeanAPAP