feat(Topology/PartitionOfUnity): add variations of partition of unity for locally compact T2 space (#12266)
add variations of `PartitionOfUnity` and Urysohn's lemma:
- In a locally compact T2 space `X`, for a compact set `t` and a finite family of open sets `{s i}` such that `t ⊆ ⋃ i, s i`, there is a family of continuous function `{f i}` supported in `s i`, `∑ i, f i x = 1` on `t` and `0 ≤ f x ≤ 1`. 
- In a locally compact regular space `X`, for a compact set `t` and a closed set `s` that are disjoint, there is a continuous function `f` with compact support which is `1` on `t` and `0` on `s` and `f x Icc (0 : ℝ) 1 ` for all `x`.

The former is formalized as a partition of unity. For this purpose, I extended `PartialRefiment` to include a predicate `p` about the refined set. With this, we can require that the refined open set has a compact closure. This can be applied when `X` is locally compact and T2.
 
These variations are needed in order to prove that [rieszContentAux](https://leanprover-community.github.io/mathlib4_docs/Mathlib/MeasureTheory/Integral/RieszMarkovKakutani.html#rieszContentAux) is indeed a Borel measure (following Rudin "Real and complex analysis"), now in #18775.



Co-authored-by: Yoh Tanimoto <57562556+yoh-tanimoto@users.noreply.github.com>