feat: the upstreamDecl linter (#18702)
This linter is intended to work much like `minImports` in helping with splitting files. It places a warning on each set of declarations that could be moved to an earlier file. It takes into account dependencies on declarations in the current file, so that a whole block of declarations can be moved even if there are dependencies between the declarations in the block.

Since definitions are often deliberately placed in their own file, this linter does not place a warning on any declaration depending on a definition in the current file.



Co-authored-by: Anne C.A. Baanen <vierkantor@vierkantor.com>
Co-authored-by: Anne Baanen <Vierkantor@users.noreply.github.com>