feat: `#parse` -- a command to parse text and log outputs (#16305)
This is a spin-off of #16230.  It introduces a function written by Matt that allows to use `#guard_msgs` to check parsing errors.

The new command is called `#guard_exceptions` and mimics very closely `#guard_msgs`.  As a way of showing how to use it, the PR also adds tests for Stacks Project Tags.



Co-authored-by: Eric Wieser <wieser.eric@gmail.com>
Co-authored-by: Matthew Ballard <matt@mrb.email>