refactor(Algebra/Category/ModuleCat): make `ModuleCat.Hom` a structure
This PR follows the approach of #19065 to solve painful definitional equalities problems by forbidding the identification of `A ⟶ B` and `A →ₗ[R] B`. These are not equal reducibly, so tactics get confused if one is replaced by the other.

Currently the PR consists mostly of me playing "type tetris" to get everything to compile again: I haven't looked for which proofs can be cleaned up as a result.

This commit is not quite finished, please see the Zulip thread for more details.