feat(Analysis/BoxIntegral): Add BoxPartition defined by subdivision of the integer lattice (#16675)
Let `n` be a positive integer. This PR defines `BoxIntegral.unitPartition.Box` which are boxes in `ι → ℝ` obtained by dividing evenly the boxes of the integer lattice`ι → ℤ` into smaller boxes of sides of length `1 / n`.  

These boxes will be used in another PR #12405 that relates point counting in sets and integration.

This PR is part of the proof of the Analytic Class Number Formula. 



Co-authored-by: Xavier Roblot <46200072+xroblot@users.noreply.github.com>