feat(NumberTheory/NumberField/Completion): API for `InfinitePlace ℚ` (#19644)
From the FLT project.

When trying to do some calculations on adeles for FLT I realised that the completion of `ℚ` at its infinite place was not equal to `ℝ` and moreover we basically knew nothing about it. After this PR we at least know which rationals are in its open unit ball (which was what I needed).