feat: specialize theorems about filters disjoint from the cocompact filter to singleton sets (#19717)
Let `X, Y` be types and let `f` be a filter on `Y` which is disjoint from the cocompact filter. We prove that `𝓝 x ×ˢ f ≤ 𝓝ˢ ({x} ×ˢ Set.univ)` for all `x : X`. This is a special case of the existing theorem `nhdsSet_prod_le_of_disjoint_cocompact`.

We also prove the analogous result with `X` and `Y` swapped.