feat(NumberTheory/NumberField/FinitePlaces): the finite places of a number field (#19667)
This file defines finite places of a number field `K` as absolute values coming from an embedding
into a completion of `K` associated to a non-zero prime ideal of `𝓞 K` with some basic APIs.

This is part of a project which aims to define heights. The next step is the product formula.