doc(RingTheory/Polynomial/HilbertPoly): fix a typo (#19833)
In this pull request, we fix a typo in the file RingTheory/Polynomial/HilbertPoly.lean: originally, the docstring at the beginning of the file said "the polynomial `(n + 1)···(n + d)/d!`". But because the polynomial we mention is in `F[X]`, we must change it to `(X + 1)···(X + d)/d!`.