chore: delete `nonempty_toType_aleph` (#19919)
This instance was a byproduct of defining [`MeasurableSpace.generateMeasurableRec`](https://leanprover-community.github.io/mathlib4_docs/Mathlib/MeasureTheory/MeasurableSpace/Card.html#MeasurableSpace.generateMeasurableRec) in terms of the very ugly `(ℵ_ o).ord.toType`. Now that we properly define it in terms of an ordinal, we have no need for this (and frankly we should not encourage it either).