refactor(Algebra/Category/ModuleCat): try to use a type synonym for `Module.compHom` instead of adding it as a local instance
This is an attempt to get rid of annoying defeq issues involving `ModuleCat.restrictScalars`: this was previously just a `def` but with a lot of its own instances. This meant it got unfolded all the time, resulting in failing instance search. Instead we try to define a `structure`.

First I tried making `Module.RestrictScalars` a `def` and inserting `out` and `into` wherever necessary to define maps, but that runs into the issue that the unfolding happens before we give a value to the maps. (Namely, it happens whenever defining a map because `ofHom` wants to unfold `ModuleCat.restrictScalars` into a `Type`.)