chore(Logic/ExistsUnique): amend usages of existsUnique to follow naming convention (#20028)
As a definition `ExistsUnique`, the spelling of usages should be `existsUnique` rather than `exists_unique`. The majority of applications already conform to this (96), but there are some (30) exceptions. This PR fixes those.

The large import is the import of Tactic.Alias to Logic.ExistsUnique, and is necessary if deprecations are to be performed here.